move to safe subdomain, tweak UI, add og images + update metadata
