+++ app/page.tsx
@@ -23,7 +23,7 @@ const subdomainToLabel = {
   [subdomains.WALLET.base]: "Wallet",
   [subdomains.ENS.base]: "ENS",
   [subdomains["7702BEAT"].base]: "7702 Beat",
-  [subdomains["712HASH"].base]: "712 Hash",
+  [subdomains.SAFE.base]: "Safe",
 };
 
 const Btn = ({
