+++ app/safe/calldata-decoder/page.tsx
@@ -1,0 +1,6 @@
+import { CalldataDecoderPage as CalldataDecoderP } from "@/components/pages/CalldataDecoderPage";
+
+const CalldataDecoderPage = () => {
+  return <CalldataDecoderP headerText="Safe Calldata Decoder" />;
+};
+export default CalldataDecoderPage;
